refactor: Navigationbar에서 로고 분리
